requirements.txt
# Core Dependencies
scrapy>=2.11.0
scrapy-playwright>=0.0.34
playwright>=1.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml>=6.0

# HTTP & Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Data Processing
pandas>=2.1.0
pydantic>=2.5.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.0

# CLI
typer>=0.9.0
rich>=13.7.0